+++ native/src/main/java/io/ballerina/stdlib/http/api/client/actions/AbstractHTTPAction.java
@@ -66,7 +66,6 @@ import java.io.OutputStream;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.nio.charset.StandardCharsets;
-import java.util.Objects;
 
 import static io.ballerina.runtime.api.constants.RuntimeConstants.BALLERINA_VERSION;
 import static io.ballerina.stdlib.http.api.HttpConstants.ANN_CONFIG_ATTR_COMPRESSION;
@@ -266,13 +265,11 @@ public abstract class AbstractHTTPAction {
         //Make the request associate with this response consumable again so that it can be reused.
         checkDirtiness(dataContext, outboundRequestMsg);
 
-        if (Objects.nonNull(dataContext.getEnvironment().getStrandLocal(HttpConstants.MAIN_STRAND))) {
-            Object sourceHandler = outboundRequestMsg.getProperty(HttpConstants.SRC_HANDLER);
-            if (sourceHandler == null) {
+        Object sourceHandler = outboundRequestMsg.getProperty(HttpConstants.SRC_HANDLER);
+        if (sourceHandler == null) {
 
-                outboundRequestMsg.setProperty(HttpConstants.SRC_HANDLER,
-                        dataContext.getEnvironment().getStrandLocal(HttpConstants.SRC_HANDLER));
-            }
+            outboundRequestMsg.setProperty(HttpConstants.SRC_HANDLER,
+                    dataContext.getEnvironment().getStrandLocal(HttpConstants.SRC_HANDLER));
         }
 
         Object poolableByteBufferFactory = outboundRequestMsg.getProperty(HttpConstants.POOLED_BYTE_BUFFER_FACTORY);