+++ native/src/main/java/io/ballerina/stdlib/http/transport/contractimpl/listener/states/http2/Expect100ContinueHeaderReceived.java
@@ -61,7 +61,7 @@ public class Expect100ContinueHeaderReceived implements ListenerState {
 
     @Override
     public void readInboundRequestBody(Http2SourceHandler http2SourceHandler, Http2DataFrame dataFrame)
-            throws Http2Exception {
+            throws Http2Exception, ServerConnectorException {
         //Client might start sending request body without waiting for the 100-continue response.
         http2MessageStateContext.setListenerState(new ReceivingEntityBody(http2MessageStateContext));
         http2MessageStateContext.getListenerState().readInboundRequestBody(http2SourceHandler, dataFrame);